TimezoneMap: fix fractional timezone offset formatting
When formatting fractional timezone offsets, force the "en_US" locale
to ensure that "." is used as the decimal separator instead of "," that
some (e.g. nordic) locales use.

This fixes the issue that depending on the value of `Intl.defaultLocale`,
the map would try to load SVGs with wrong file names such as "tz_5,5.svg"
instead of "tz_5.5.svg" for India.