Optimize loadLocalizedLanguages()
The languages are sorted by names with diacritics removed to ensure
that e.g. "Íslenska" ("Islenska") appears next to "Italiano" in the
list of languages, not somewhere at the end of the list.

Before, the function collected all localized languages into a list, and
then sorted the list with a comparator function that removed diacritics
from the localized language names on the fly at each comparison.

Now, the function makes use of an ordered SplayTreeMap using non-
diacritic language names as keys and returns a lazy iterable of values
in the right order. This ensures that diacritics are only removed once
per language name, and the final collection of languages is ordered by
a binary search at insertion time instead of doing a quicksort of an
unordered list at the end.