refactor: revise and rename setupAppLocalizations()
The old name was misleading because the function doesn't do anything
with `AppLocalizations` but initializes `Intl.defaultLocale`. Moreover,
there needs to be a way to pass another locale to make sure dates,
times, and numbers are formatted according to the currently selected
locale instead of sticking to the system locale.