chore: re-add generated files to the repository (#480)
We've previously removed all generated files from this repo in #454 and
added a hook that's supposed to run before publishing packages. It looks
like this currently doesn't work correctly, though -
`ubuntu_localizations` `0.5.2` misses the generated translation files.
I think the easiest solution is to revert to the previous state again
and check in all files that need to be included in the published
package.